+++ mobile/app/_layout.tsx
@@ -1,0 +1,45 @@
+import { BaiJamjuree_700Bold } from '@expo-google-fonts/bai-jamjuree'
+import {
+  Roboto_400Regular,
+  Roboto_700Bold,
+  useFonts,
+} from '@expo-google-fonts/roboto'
+import { SplashScreen, Stack } from 'expo-router'
+import { StatusBar } from 'expo-status-bar'
+import { styled } from 'nativewind'
+import React from 'react'
+import { ImageBackground } from 'react-native'
+import blurBg from '../src/assets/bg-blur.png'
+import Stripes from '../src/assets/stripes.svg'
+
+const StyledStripes = styled(Stripes)
+
+export default function Layout() {
+  const [hasLoadedFonts] = useFonts({
+    Roboto_400Regular,
+    Roboto_700Bold,
+    BaiJamjuree_700Bold,
+  })
+
+  if (!hasLoadedFonts) {
+    return <SplashScreen />
+  }
+
+  return (
+    <ImageBackground
+      source={blurBg}
+      className="relative flex-1 bg-gray-900"
+      imageStyle={{ position: 'absolute', left: '-100%' }}
+    >
+      <StyledStripes className="absolute left-2" />
+      <StatusBar style="light" translucent />
+
+      <Stack
+        screenOptions={{
+          headerShown: false,
+          contentStyle: { backgroundColor: 'trasnparent' },
+        }}
+      />
+    </ImageBackground>
+  )
+}
